requirements.txt
boto==2.6.0
django==1.8.18
djangorestframework==3.6.3
elementtree==1.2.7-20070827-preview
html5lib==0.90
markdown==2.0.3
mysql-python==1.2.3
psycopg2==2.6
pysqlite==2.6.3
python-openid==2.2.5
south==0.7.6
virtualenv==1.11.2
virtualenvwrapper-win==1.1.5
